Add support for HA PROXY v1
Resolves https://github.com/hannesm/tlstunnel/issues/6

This commit adds a flag (`--haproxy1`) which causes tlstunnel to send
connection details to the destination upon connection.

This is useful if running `nginx`, `Varnish` or similar behind `tlstunnel` in order
to obtain the IP of the client.

Example of a TCP/IPv4 connection from 127.0.0.1:39837 to 127.0.0.1:4433:
```
PROXY TCP4 127.0.0.1 127.0.0.1 39837 4433
```

Relevant `nginx` documentation on how to use the PROXY protocol: https://www.nginx.com/resources/admin-guide/proxy-protocol/