new feature: client authentication
an optional command line argument was introduced, named "--auth", which
receives a filename that contains a list of pem-encoded certificates.

upon successfully establishing a TLS connection on the frontend, the server
immediately renegotiates the TLS session, this time asking for a client
certificate (and providing the acceptable_cas, which are the subjects of the
provided CAs).

the reason for renegotiation, which adds more roundtrips, is to send the client
certificate only over a (server-)authenticated encrypted channel, so
eavesdroppers are not able to track the actual client

logging now includes information about the peer certificate (subject and serial
number) if present.  after renegotiation, another log entry is generated.

when --auth is not provided, nothing changes (famous last words)